Crate collections

Deprecated since 1.20.0
: collections moved to alloc

🔬 This is a nightly-only experimental API. (collections #27783)

this library is unlikely to be stabilized in its current form or name

Collection types

See std::collections for a detailed discussion of collections in Rust.

Reexports

pub use alloc::binary_heap::BinaryHeap;

pub use alloc::btree_map::BTreeMap;

pub use alloc::btree_set::BTreeSet;

pub use alloc::linked_list::LinkedList;

pub use alloc::vec_deque::VecDeque;

pub use alloc::string::String;

pub use alloc::vec::Vec;

Modules

binary_heap
A priority queue implemented with a binary heap.

borrow
A module for working with borrowed data.

btree_map
A map based on a B-Tree.

btree_set
A set based on a B-Tree.

fmt
Utilities for formatting and printing Strings

linked_list
A doubly-linked list with owned nodes.

slice
A dynamically-sized view into a contiguous sequence, [T].

str
Unicode string slices.

string
A UTF-8 encoded, growable string.

vec
A contiguous growable array type with heap-allocated contents, written Vec<T>.

vec_deque
A double-ended queue implemented with a growable ring buffer.

range 	 [
Experimental
]
Range syntax.

Macros

format
Use the syntax described in std::fmt to create a value of type String. See std::fmt for more information.

vec
Creates a Vec containing the arguments.

Enums

Bound
An endpoint of a range of keys.