Struct collections::binary_heap::BinaryHeap

pub struct BinaryHeap<T> { /* fields omitted */ }

A priority queue implemented with a binary heap.

This will be a max-heap.

It is a logic error for an item to be modified in such a way that the item's ordering relative to any other item, as determined by the Ord trait, changes while it is in the heap. This is normally only possible through Cell, RefCell, global state, I/O, or unsafe code.

Examples

use std::collections::BinaryHeap;

// Type inference lets us omit an explicit type signature (which
// would be `BinaryHeap<i32>` in this example).
let mut heap = BinaryHeap::new();

// We can use peek to look at the next item in the heap. In this case,
// there's no items in there yet so we get None.
assert_eq!(heap.peek(), None);

// Let's add some scores...
heap.push(1);
heap.push(5);
heap.push(2);

// Now peek shows the most important item in the heap.
assert_eq!(heap.peek(), Some(&5));

// We can check the length of a heap.
assert_eq!(heap.len(), 3);

// We can iterate over the items in the heap, although they are returned in
// a random order.
for x in &heap {
    println!("{}", x);
}

// If we instead pop these scores, they should come back in order.
assert_eq!(heap.pop(), Some(5));
assert_eq!(heap.pop(), Some(2));
assert_eq!(heap.pop(), Some(1));
assert_eq!(heap.pop(), None);

// We can clear the heap of any remaining items.
heap.clear();

// The heap should now be empty.
assert!(heap.is_empty())Run

Methods

impl<T: Ord> BinaryHeap<T>

fn new() -> BinaryHeap<T>

Creates an empty BinaryHeap as a max-heap.

Examples

Basic usage:

use std::collections::BinaryHeap;
let mut heap = BinaryHeap::new();
heap.push(4);Run

fn with_capacity(capacity: usize) -> BinaryHeap<T>

Creates an empty BinaryHeap with a specific capacity. This preallocates enough memory for capacity elements, so that the BinaryHeap does not have to be reallocated until it contains at least that many values.

Examples

Basic usage:

use std::collections::BinaryHeap;
let mut heap = BinaryHeap::with_capacity(10);
heap.push(4);Run

fn iter(&self) -> Iter<T>

Returns an iterator visiting all values in the underlying vector, in arbitrary order.

Examples

Basic usage:

use std::collections::BinaryHeap;
let heap = BinaryHeap::from(vec![1, 2, 3, 4]);

// Print 1, 2, 3, 4 in arbitrary order
for x in heap.iter() {
    println!("{}", x);
}Run

fn peek(&self) -> Option<&T>

Returns the greatest item in the binary heap, or None if it is empty.

Examples

Basic usage:

use std::collections::BinaryHeap;
let mut heap = BinaryHeap::new();
assert_eq!(heap.peek(), None);

heap.push(1);
heap.push(5);
heap.push(2);
assert_eq!(heap.peek(), Some(&5));
Run

fn peek_mut(&mut self) -> Option<PeekMut<T>>

Returns a mutable reference to the greatest item in the binary heap, or None if it is empty.

Note: If the PeekMut value is leaked, the heap may be in an inconsistent state.

Examples

Basic usage:

use std::collections::BinaryHeap;
let mut heap = BinaryHeap::new();
assert!(heap.peek_mut().is_none());

heap.push(1);
heap.push(5);
heap.push(2);
{
    let mut val = heap.peek_mut().unwrap();
    *val = 0;
}
assert_eq!(heap.peek(), Some(&2));Run

fn capacity(&self) -> usize

Returns the number of elements the binary heap can hold without reallocating.

Examples

Basic usage:

use std::collections::BinaryHeap;
let mut heap = BinaryHeap::with_capacity(100);
assert!(heap.capacity() >= 100);
heap.push(4);Run

fn reserve_exact(&mut self, additional: usize)

Reserves the minimum capacity for exactly additional more elements to be inserted in the given BinaryHeap. Does nothing if the capacity is already sufficient.

Note that the allocator may give the collection more space than it requests. Therefore capacity can not be relied upon to be precisely minimal. Prefer reserve if future insertions are expected.

Panics

Panics if the new capacity overflows usize.

Examples

Basic usage:

use std::collections::BinaryHeap;
let mut heap = BinaryHeap::new();
heap.reserve_exact(100);
assert!(heap.capacity() >= 100);
heap.push(4);Run

fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more elements to be inserted in the BinaryHeap. The collection may reserve more space to avoid frequent reallocations.

Panics

Panics if the new capacity overflows usize.

Examples

Basic usage:

use std::collections::BinaryHeap;
let mut heap = BinaryHeap::new();
heap.reserve(100);
assert!(heap.capacity() >= 100);
heap.push(4);Run

fn shrink_to_fit(&mut self)

Discards as much additional capacity as possible.

Examples

Basic usage:

use std::collections::BinaryHeap;
let mut heap: BinaryHeap<i32> = BinaryHeap::with_capacity(100);

assert!(heap.capacity() >= 100);
heap.shrink_to_fit();
assert!(heap.capacity() == 0);Run

fn pop(&mut self) -> Option<T>

Removes the greatest item from the binary heap and returns it, or None if it is empty.

Examples

Basic usage:

use std::collections::BinaryHeap;
let mut heap = BinaryHeap::from(vec![1, 3]);

assert_eq!(heap.pop(), Some(3));
assert_eq!(heap.pop(), Some(1));
assert_eq!(heap.pop(), None);Run

fn push(&mut self, item: T)

Pushes an item onto the binary heap.

Examples

Basic usage:

use std::collections::BinaryHeap;
let mut heap = BinaryHeap::new();
heap.push(3);
heap.push(5);
heap.push(1);

assert_eq!(heap.len(), 3);
assert_eq!(heap.peek(), Some(&5));Run

fn push_pop(&mut self, item: T) -> T

Deprecated since 1.13.0
: use peek_mut instead

Unstable (binary_heap_extras #28147)
: needs to be audited

Pushes an item onto the binary heap, then pops the greatest item off the queue in an optimized fashion.

Examples

Basic usage:

#![feature(binary_heap_extras)]
#![allow(deprecated)]

use std::collections::BinaryHeap;
let mut heap = BinaryHeap::new();
heap.push(1);
heap.push(5);

assert_eq!(heap.push_pop(3), 5);
assert_eq!(heap.push_pop(9), 9);
assert_eq!(heap.len(), 2);
assert_eq!(heap.peek(), Some(&3));Run

fn replace(&mut self, item: T) -> Option<T>

Deprecated since 1.13.0
: use peek_mut instead

Unstable (binary_heap_extras #28147)
: needs to be audited

Pops the greatest item off the binary heap, then pushes an item onto the queue in an optimized fashion. The push is done regardless of whether the binary heap was empty.

Examples

Basic usage:

#![feature(binary_heap_extras)]
#![allow(deprecated)]

use std::collections::BinaryHeap;
let mut heap = BinaryHeap::new();

assert_eq!(heap.replace(1), None);
assert_eq!(heap.replace(3), Some(1));
assert_eq!(heap.len(), 1);
assert_eq!(heap.peek(), Some(&3));Run

fn into_vec(self) -> Vec<T>

Consumes the BinaryHeap and returns the underlying vector in arbitrary order.

Examples

Basic usage:

use std::collections::BinaryHeap;
let heap = BinaryHeap::from(vec![1, 2, 3, 4, 5, 6, 7]);
let vec = heap.into_vec();

// Will print in some order
for x in vec {
    println!("{}", x);
}Run

fn into_sorted_vec(self) -> Vec<T>

Consumes the BinaryHeap and returns a vector in sorted (ascending) order.

Examples

Basic usage:

use std::collections::BinaryHeap;

let mut heap = BinaryHeap::from(vec![1, 2, 4, 5, 7]);
heap.push(6);
heap.push(3);

let vec = heap.into_sorted_vec();
assert_eq!(vec, [1, 2, 3, 4, 5, 6, 7]);Run

fn len(&self) -> usize

Returns the length of the binary heap.

Examples

Basic usage:

use std::collections::BinaryHeap;
let heap = BinaryHeap::from(vec![1, 3]);

assert_eq!(heap.len(), 2);Run

fn is_empty(&self) -> bool

Checks if the binary heap is empty.

Examples

Basic usage:

use std::collections::BinaryHeap;
let mut heap = BinaryHeap::new();

assert!(heap.is_empty());

heap.push(3);
heap.push(5);
heap.push(1);

assert!(!heap.is_empty());Run

fn drain(&mut self) -> Drain<T>

Clears the binary heap, returning an iterator over the removed elements.

The elements are removed in arbitrary order.

Examples

Basic usage:

use std::collections::BinaryHeap;
let mut heap = BinaryHeap::from(vec![1, 3]);

assert!(!heap.is_empty());

for x in heap.drain() {
    println!("{}", x);
}

assert!(heap.is_empty());Run

fn clear(&mut self)

Drops all items from the binary heap.

Examples

Basic usage:

use std::collections::BinaryHeap;
let mut heap = BinaryHeap::from(vec![1, 3]);

assert!(!heap.is_empty());

heap.clear();

assert!(heap.is_empty());Run

fn append(&mut self, other: &mut Self)

Moves all the elements of other into self, leaving other empty.

Examples

Basic usage:

use std::collections::BinaryHeap;

let v = vec![-10, 1, 2, 3, 3];
let mut a = BinaryHeap::from(v);

let v = vec![-20, 5, 43];
let mut b = BinaryHeap::from(v);

a.append(&mut b);

assert_eq!(a.into_sorted_vec(), [-20, -10, 1, 2, 3, 3, 5, 43]);
assert!(b.is_empty());Run