Module collections::btree::map

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

Reexports

use core::prelude::v1::*;

use core::cmp::Ordering;

use core::fmt::Debug;

use core::hash::{Hash, Hasher};

use core::iter::FromIterator;

use core::marker::PhantomData;

use core::ops::Index;

use core::{fmt, intrinsics, mem, ptr};

use borrow::Borrow;

use Bound::{self, Included, Excluded, Unbounded};

use super::node::{self, NodeRef, Handle, marker};

use super::search;

use super::node::InsertResult::*;

use super::node::ForceResult::*;

use super::search::SearchResult::*;

use self::UnderflowResult::*;

use self::Entry::*;

Structs

BTreeMap
A map based on a B-Tree.

IntoIter
An owning iterator over a BTreeMap's entries.

Iter
An iterator over a BTreeMap's entries.

IterMut
A mutable iterator over a BTreeMap's entries.

Keys
An iterator over a BTreeMap's keys.

OccupiedEntry
An occupied Entry.

VacantEntry
A vacant Entry.

Values
An iterator over a BTreeMap's values.

Range 	 [Unstable]
An iterator over a sub-range of BTreeMap's entries.

RangeMut 	 [Unstable]
A mutable iterator over a sub-range of BTreeMap's entries.

Enums

Entry
A view into a single entry in a map, which may either be vacant or occupied.

UnderflowResult 	 [Unstable]

Functions

first_leaf_edge 	 [Unstable]

handle_underfull_node 	 [Unstable]

last_leaf_edge 	 [Unstable]

unwrap_unchecked 	 [Unstable]