Module collections::btree::node

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

Reexports

use core::prelude::v1::*;

use alloc::heap;

use core::marker::PhantomData;

use core::mem;

use core::nonzero::NonZero;

use core::ptr::{self, Unique};

use core::slice;

use boxed::Box;

Modules

marker 	 [Unstable]

Structs

BoxedNode 	 [Unstable]

Handle 	 [Unstable]

InternalNode 	 [Unstable]

LeafNode 	 [Unstable]

NodeRef 	 [Unstable]
A reference to a node.

Root 	 [Unstable]
An owned tree. Note that despite being owned, this does not have a destructor, and must be cleaned up manually.

Enums

ForceResult 	 [Unstable]

InsertResult 	 [Unstable]

Constants

B 	 [Unstable]

CAPACITY 	 [Unstable]

Functions

slice_insert 	 [Unstable]

slice_remove 	 [Unstable]