Struct collections::btree::node::Handle

pub struct Handle<Node, Type> {
    node: Node,
    idx: usize,
    _marker: PhantomData<Type>,
}

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

Fields

node
Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

idx
Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

_marker
Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

Methods

impl<Node, Type> Handle<Node, Type>

fn into_node(self) -> Node

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

impl<BorrowType, K, V, NodeType> Handle<NodeRef<BorrowType, K, V, NodeType>, KV>

fn new_kv(node: NodeRef<BorrowType, K, V, NodeType>, idx: usize) -> Self

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

fn left_edge(self) -> Handle<NodeRef<BorrowType, K, V, NodeType>, Edge>

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

fn right_edge(self) -> Handle<NodeRef<BorrowType, K, V, NodeType>, Edge>

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

impl<BorrowType, K, V, NodeType, HandleType> Handle<NodeRef<BorrowType, K, V, NodeType>, HandleType>

fn reborrow(&self) -> Handle<NodeRef<Immut, K, V, NodeType>, HandleType>

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

impl<'a, K, V, NodeType, HandleType> Handle<NodeRef<Mut<'a>, K, V, NodeType>, HandleType>

unsafe fn reborrow_mut(&mut self) -> Handle<NodeRef<Mut, K, V, NodeType>, HandleType>

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

impl<BorrowType, K, V, NodeType> Handle<NodeRef<BorrowType, K, V, NodeType>, Edge>

fn new_edge(node: NodeRef<BorrowType, K, V, NodeType>, idx: usize) -> Self

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

fn left_kv(self) -> Result<Handle<NodeRef<BorrowType, K, V, NodeType>, KV>, Self>

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

fn right_kv(self) -> Result<Handle<NodeRef<BorrowType, K, V, NodeType>, KV>, Self>

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

impl<'a, K, V> Handle<NodeRef<Mut<'a>, K, V, Leaf>, Edge>

fn insert_fit(&mut self, key: K, val: V) -> *mut V

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

fn insert(self, key: K, val: V) -> (InsertResult<'a, K, V, Leaf>, *mut V)

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

impl<'a, K, V> Handle<NodeRef<Mut<'a>, K, V, Internal>, Edge>

fn correct_parent_link(self)

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

unsafe fn cast_unchecked<NewType>(&mut self) -> Handle<NodeRef<Mut, K, V, NewType>, Edge>

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

fn insert_fit(&mut self, key: K, val: V, edge: Root<K, V>)

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

fn insert(self, key: K, val: V, edge: Root<K, V>) -> InsertResult<'a, K, V, Internal>

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

impl<BorrowType, K, V> Handle<NodeRef<BorrowType, K, V, Internal>, Edge>

fn descend(self) -> NodeRef<BorrowType, K, V, LeafOrInternal>

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

impl<'a, K: 'a, V: 'a, NodeType> Handle<NodeRef<Immut<'a>, K, V, NodeType>, KV>

fn into_kv(self) -> (&'a K, &'a V)

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

impl<'a, K: 'a, V: 'a, NodeType> Handle<NodeRef<Mut<'a>, K, V, NodeType>, KV>

fn into_kv_mut(self) -> (&'a mut K, &'a mut V)

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

impl<'a, K, V, NodeType> Handle<NodeRef<Mut<'a>, K, V, NodeType>, KV>

fn kv_mut(&mut self) -> (&mut K, &mut V)

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

impl<'a, K, V> Handle<NodeRef<Mut<'a>, K, V, Leaf>, KV>

fn split(self) -> (NodeRef<Mut<'a>, K, V, Leaf>, K, V, Root<K, V>)

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

fn remove(self) -> (Handle<NodeRef<Mut<'a>, K, V, Leaf>, Edge>, K, V)

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

impl<'a, K, V> Handle<NodeRef<Mut<'a>, K, V, Internal>, KV>

fn split(self) -> (NodeRef<Mut<'a>, K, V, Internal>, K, V, Root<K, V>)

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

fn can_merge(&self) -> bool

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

fn merge(self) -> Handle<NodeRef<Mut<'a>, K, V, Internal>, Edge>

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

impl<BorrowType, K, V, HandleType> Handle<NodeRef<BorrowType, K, V, LeafOrInternal>, HandleType>

fn force(self) -> ForceResult<Handle<NodeRef<BorrowType, K, V, Leaf>, HandleType>, Handle<NodeRef<BorrowType, K, V, Internal>, HandleType>>

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead