Trait collections::btree::Recover

trait Recover<Q: ?Sized> {
    type Key;
    fn get(&self, key: &Q) -> Option<&Self::Key>;
    fn take(&mut self, key: &Q) -> Option<Self::Key>;
    fn replace(&mut self, key: Self::Key) -> Option<Self::Key>;
}

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

Associated Types

type Key

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

Required Methods

fn get(&self, key: &Q) -> Option<&Self::Key>

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

fn take(&mut self, key: &Q) -> Option<Self::Key>

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

fn replace(&mut self, key: Self::Key) -> Option<Self::Key>

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead