Enum collections::btree_map::Entry

pub enum Entry<'a, K, V> where
    K: 'a,
    V: 'a,  {
    Vacant(VacantEntry<'a, K, V>),
    Occupied(OccupiedEntry<'a, K, V>),
}

A view into a single entry in a map, which may either be vacant or occupied.

This enum is constructed from the entry method on BTreeMap.

Variants
Vacant(VacantEntry<'a, K, V>)

A vacant entry.

Occupied(OccupiedEntry<'a, K, V>)

An occupied entry.

Methods

impl<'a, K, V> Entry<'a, K, V> where
K: Ord,

fn or_insert(self, default: V) -> &'a mut V

Ensures a value is in the entry by inserting the default if empty, and returns a mutable reference to the value in the entry.

Examples

use std::collections::BTreeMap;

let mut map: BTreeMap<&str, usize> = BTreeMap::new();
map.entry("poneyland").or_insert(12);

assert_eq!(map["poneyland"], 12);

fn or_insert_with<F>(self, default: F) -> &'a mut V where
F: FnOnce() -> V,

Ensures a value is in the entry by inserting the result of the default function if empty, and returns a mutable reference to the value in the entry.

Examples

use std::collections::BTreeMap;

let mut map: BTreeMap<&str, String> = BTreeMap::new();
let s = "hoho".to_string();

map.entry("poneyland").or_insert_with(|| s);

assert_eq!(map["poneyland"], "hoho".to_string());

fn key(&self) -> &K

Returns a reference to this entry's key.

Examples

use std::collections::BTreeMap;

let mut map: BTreeMap<&str, usize> = BTreeMap::new();
assert_eq!(map.entry("poneyland").key(), &"poneyland");

fn and_modify<F>(self, f: F) -> Entry<'a, K, V> where
F: FnMut(&mut V) -> (),

🔬 This is a nightly-only experimental API. (entry_and_modify #44733)

Provides in-place mutable access to an occupied entry before any potential inserts into the map.

Examples

#![feature(entry_and_modify)]
use std::collections::BTreeMap;

let mut map: BTreeMap<&str, usize> = BTreeMap::new();

map.entry("poneyland")
   .and_modify(|e| { *e += 1 })
   .or_insert(42);
assert_eq!(map["poneyland"], 42);

map.entry("poneyland")
   .and_modify(|e| { *e += 1 })
   .or_insert(42);
assert_eq!(map["poneyland"], 43);

impl<'a, K, V> Entry<'a, K, V> where
K: Ord,
V: Default,

fn or_default(self) -> &'a mut V

🔬 This is a nightly-only experimental API. (entry_or_default #44324)

Ensures a value is in the entry by inserting the default value if empty, and returns a mutable reference to the value in the entry.

Examples

#![feature(entry_or_default)]
use std::collections::BTreeMap;

let mut map: BTreeMap<&str, Option<usize>> = BTreeMap::new();
map.entry("poneyland").or_default();

assert_eq!(map["poneyland"], None);