Module collections::btree_setStable

Structs

BTreeSet
A set based on a B-Tree.

Difference
A lazy iterator producing elements in the set difference (in-order).

Intersection
A lazy iterator producing elements in the set intersection (in-order).

IntoIter
An owning iterator over a BTreeSet's items.

Iter
An iterator over a BTreeSet's items.

Range
An iterator over a sub-range of BTreeSet's items.

SymmetricDifference
A lazy iterator producing elements in the set symmetric difference (in-order).

Union
A lazy iterator producing elements in the set union (in-order).