Module collections::enum_setUnstable

A structure for holding a set of enum variants.

This module defines a container which uses an efficient bit mask representation to hold C-like enum variants.

Structs

EnumSet
A specialized set implementation to use enum types.

Iter
An iterator over an EnumSet

Traits

CLike
An interface for casting C-like enum to usize and back. A typically implementation is as below.