Struct collections::fmt::DebugTuple

#[must_use]

pub struct DebugTuple<'a, 'b> where
    'b: 'a,  { /* fields omitted */ }

A struct to help with fmt::Debug implementations.

This is useful when you wish to output a formatted tuple as a part of your Debug::fmt implementation.

This can be constructed by the Formatter::debug_tuple method.

Examples

use std::fmt;

struct Foo(i32, String);

impl fmt::Debug for Foo {
    fn fmt(&self, fmt: &mut fmt::Formatter) -> fmt::Result {
        fmt.debug_tuple("Foo")
           .field(&self.0)
           .field(&self.1)
           .finish()
    }
}

// prints "Foo(10, "Hello World")"
println!("{:?}", Foo(10, "Hello World".to_string()));

Methods

impl<'a, 'b> DebugTuple<'a, 'b> where
'b: 'a,

fn field(&mut self, value: &Debug) -> &mut DebugTuple<'a, 'b>

Adds a new field to the generated tuple struct output.

fn finish(&mut self) -> Result<(), Error>

Finishes output and returns any error encountered.