Struct collections::linked_list::Node

struct Node<T> {
    next: Link<T>,
    prev: Rawlink<Node<T>>,
    value: T,
}

Fields

next

prev

value

Methods

impl<T> Node<T>

fn new(v: T) -> Node<T>

fn set_next(&mut self, next: Box<Node<T>>)

Update the prev link on next, then set self's next pointer.

self.next should be None when you call this (otherwise a Node is probably being dropped by mistake).