Struct collections::slice::ChunksMut

pub struct ChunksMut<'a, T> where
    T: 'a,  { /* fields omitted */ }

An iterator over a slice in (non-overlapping) mutable chunks (size elements at a time). When the slice len is not evenly divided by the chunk size, the last slice of the iteration will be the remainder.

This struct is created by the chunks_mut method on slices.