Struct collections::slice::SplitNMut

pub struct SplitNMut<'a, T, P> where
    P: FnMut(&T) -> bool,
    T: 'a,  { /* fields omitted */ }

An iterator over subslices separated by elements that match a predicate function, limited to a given number of splits.

This struct is created by the splitn_mut method on slices.