Trait collections::slice::SliceConcatExt

pub trait SliceConcatExt<T: ?Sized> {
    type Output;
    fn concat(&self) -> Self::Output;
    fn join(&self, sep: &T) -> Self::Output;
    fn connect(&self, sep: &T) -> Self::Output;
}

🔬 This is a nightly-only experimental API. (slice_concat_ext #27747)

trait should not have to exist

An extension trait for concatenating slices

Associated Types

type Output

🔬 This is a nightly-only experimental API. (slice_concat_ext #27747)

trait should not have to exist

The resulting type after concatenation

Required Methods

fn concat(&self) -> Self::Output

Flattens a slice of T into a single value Self::Output.

Examples

assert_eq!(["hello", "world"].concat(), "helloworld");Run

fn join(&self, sep: &T) -> Self::Output

Flattens a slice of T into a single value Self::Output, placing a given separator between each.

Examples

assert_eq!(["hello", "world"].join(" "), "hello world");Run

fn connect(&self, sep: &T) -> Self::Output

Deprecated since 1.3.0
: renamed to join