Trait collections::slice::SliceExtStable

pub trait SliceExt {
    type Item;

    fn sort_by<F>(&mut self, compare: F) where F: FnMut(&Self::Item, &Self::Item) -> Ordering;
    fn move_from(&mut self, src: Vec<Self::Item>, start: usize, end: usize) -> usize;
    fn slice(&self, start: usize, end: usize) -> &[Self::Item];
    fn slice_from(&self, start: usize) -> &[Self::Item];
    fn slice_to(&self, end: usize) -> &[Self::Item];
    fn split_at(&self, mid: usize) -> (&[Self::Item], &[Self::Item]);
    fn iter(&self) -> Iter<Self::Item>;
    fn split<F>(&self, pred: F) -> Split<Self::Item, F> where F: FnMut(&Self::Item) -> bool;
    fn splitn<F>(&self, n: usize, pred: F) -> SplitN<Self::Item, F> where F: FnMut(&Self::Item) -> bool;
    fn rsplitn<F>(&self, n: usize, pred: F) -> RSplitN<Self::Item, F> where F: FnMut(&Self::Item) -> bool;
    fn windows(&self, size: usize) -> Windows<Self::Item>;
    fn chunks(&self, size: usize) -> Chunks<Self::Item>;
    fn get(&self, index: usize) -> Option<&Self::Item>;
    fn first(&self) -> Option<&Self::Item>;
    fn tail(&self) -> &[Self::Item];
    fn init(&self) -> &[Self::Item];
    fn last(&self) -> Option<&Self::Item>;
    unsafe fn get_unchecked(&self, index: usize) -> &Self::Item;
    fn as_ptr(&self) -> *const Self::Item;
    fn binary_search_by<F>(&self, f: F) -> Result<usize, usize> where F: FnMut(&Self::Item) -> Ordering;
    fn len(&self) -> usize;
    fn get_mut(&mut self, index: usize) -> Option<&mut Self::Item>;
    fn as_mut_slice(&mut self) -> &mut [Self::Item];
    fn slice_mut(&mut self, start: usize, end: usize) -> &mut [Self::Item];
    fn slice_from_mut(&mut self, start: usize) -> &mut [Self::Item];
    fn slice_to_mut(&mut self, end: usize) -> &mut [Self::Item];
    fn iter_mut(&mut self) -> IterMut<Self::Item>;
    fn first_mut(&mut self) -> Option<&mut Self::Item>;
    fn tail_mut(&mut self) -> &mut [Self::Item];
    fn init_mut(&mut self) -> &mut [Self::Item];
    fn last_mut(&mut self) -> Option<&mut Self::Item>;
    fn split_mut<F>(&mut self, pred: F) -> SplitMut<Self::Item, F> where F: FnMut(&Self::Item) -> bool;
    fn splitn_mut<F>(&mut self, n: usize, pred: F) -> SplitNMut<Self::Item, F> where F: FnMut(&Self::Item) -> bool;
    fn rsplitn_mut<F>(&mut self, n: usize, pred: F) -> RSplitNMut<Self::Item, F> where F: FnMut(&Self::Item) -> bool;
    fn chunks_mut(&mut self, chunk_size: usize) -> ChunksMut<Self::Item>;
    fn swap(&mut self, a: usize, b: usize);
    fn split_at_mut(&mut self, mid: usize) -> (&mut [Self::Item], &mut [Self::Item]);
    fn reverse(&mut self);
    unsafe fn get_unchecked_mut(&mut self, index: usize) -> &mut Self::Item;
    fn as_mut_ptr(&mut self) -> *mut Self::Item;
    fn to_vec(&self) -> Vec<Self::Item> where Self::Item: Clone;
    fn permutations(&self) -> Permutations<Self::Item> where Self::Item: Clone;
    fn clone_from_slice(&mut self, &[Self::Item]) -> usize where Self::Item: Clone;
    fn sort(&mut self) where Self::Item: Ord;
    fn binary_search(&self, x: &Self::Item) -> Result<usize, usize> where Self::Item: Ord;
    fn next_permutation(&mut self) -> bool where Self::Item: Ord;
    fn prev_permutation(&mut self) -> bool where Self::Item: Ord;
    fn position_elem(&self, t: &Self::Item) -> Option<usize> where Self::Item: PartialEq;
    fn rposition_elem(&self, t: &Self::Item) -> Option<usize> where Self::Item: PartialEq;
    fn contains(&self, x: &Self::Item) -> bool where Self::Item: PartialEq;
    fn starts_with(&self, needle: &[Self::Item]) -> bool where Self::Item: PartialEq;
    fn ends_with(&self, needle: &[Self::Item]) -> bool where Self::Item: PartialEq;
    fn into_vec(self: Box<Self>) -> Vec<Self::Item>;

    fn is_empty(&self) -> bool { ... }
    fn binary_search_elem(&self, x: &Self::Item) -> Result<usize, usize> where Self::Item: Ord { ... }
}

Allocating extension methods for slices.

Associated Types

type Item

Required Methods

fn sort_by<F>(&mut self, compare: F) where F: FnMut(&Self::Item, &Self::Item) -> Ordering

Sorts the slice, in place, using compare to compare elements.

This sort is O(n log n) worst-case and stable, but allocates approximately 2 * n, where n is the length of self.

Examples
fn main() { let mut v = [5, 4, 1, 3, 2]; v.sort_by(|a, b| a.cmp(b)); assert!(v == [1, 2, 3, 4, 5]); // reverse sorting v.sort_by(|a, b| b.cmp(a)); assert!(v == [5, 4, 3, 2, 1]); }

let mut v = [5, 4, 1, 3, 2];
v.sort_by(|a, b| a.cmp(b));
assert!(v == [1, 2, 3, 4, 5]);

// reverse sorting
v.sort_by(|a, b| b.cmp(a));
assert!(v == [5, 4, 3, 2, 1]);

fn move_from(&mut self, src: Vec<Self::Item>, start: usize, end: usize) -> usize

Consumes src and moves as many elements as it can into self from the range [start,end).

Returns the number of elements copied (the shorter of self.len() and end - start).

Arguments

src - A mutable vector of T

start - The index into src to start copying from

end - The index into src to stop copying from

Examples
fn main() { let mut a = [1, 2, 3, 4, 5]; let b = vec![6, 7, 8]; let num_moved = a.move_from(b, 0, 3); assert_eq!(num_moved, 3); assert!(a == [6, 7, 8, 4, 5]); }

let mut a = [1, 2, 3, 4, 5];
let b = vec![6, 7, 8];
let num_moved = a.move_from(b, 0, 3);
assert_eq!(num_moved, 3);
assert!(a == [6, 7, 8, 4, 5]);

fn slice(&self, start: usize, end: usize) -> &[Self::Item]

Deprecated: use &s[start .. end] notation instead.

fn slice_from(&self, start: usize) -> &[Self::Item]

Deprecated: use &s[start..] notation instead.

fn slice_to(&self, end: usize) -> &[Self::Item]

Deprecated: use &s[..end] notation instead.

fn split_at(&self, mid: usize) -> (&[Self::Item], &[Self::Item])

Divides one slice into two at an index.

The first will contain all indices from [0, mid) (excluding the index mid itself) and the second will contain all indices from [mid, len) (excluding the index len itself).

Panics if mid > len.

Examples
fn main() { let v = [10, 40, 30, 20, 50]; let (v1, v2) = v.split_at(2); assert_eq!([10, 40], v1); assert_eq!([30, 20, 50], v2); }

let v = [10, 40, 30, 20, 50];
let (v1, v2) = v.split_at(2);
assert_eq!([10, 40], v1);
assert_eq!([30, 20, 50], v2);

fn iter(&self) -> Iter<Self::Item>

Returns an iterator over the slice.

fn split<F>(&self, pred: F) -> Split<Self::Item, F> where F: FnMut(&Self::Item) -> bool

Returns an iterator over subslices separated by elements that match pred. The matched element is not contained in the subslices.

Examples

Print the slice split by numbers divisible by 3 (i.e. [10, 40], [20], [50]):
fn main() { let v = [10, 40, 30, 20, 60, 50]; for group in v.split(|num| *num % 3 == 0) { println!("{:?}", group); } }

let v = [10, 40, 30, 20, 60, 50];
for group in v.split(|num| *num % 3 == 0) {
    println!("{:?}", group);
}

fn splitn<F>(&self, n: usize, pred: F) -> SplitN<Self::Item, F> where F: FnMut(&Self::Item) -> bool

Returns an iterator over subslices separated by elements that match pred, limited to splitting at most n times. The matched element is not contained in the subslices.

Examples

Print the slice split once by numbers divisible by 3 (i.e. [10, 40], [20, 60, 50]):
fn main() { let v = [10, 40, 30, 20, 60, 50]; for group in v.splitn(1, |num| *num % 3 == 0) { println!("{:?}", group); } }

let v = [10, 40, 30, 20, 60, 50];
for group in v.splitn(1, |num| *num % 3 == 0) {
    println!("{:?}", group);
}

fn rsplitn<F>(&self, n: usize, pred: F) -> RSplitN<Self::Item, F> where F: FnMut(&Self::Item) -> bool

Returns an iterator over subslices separated by elements that match pred limited to splitting at most n times. This starts at the end of the slice and works backwards. The matched element is not contained in the subslices.

Examples

Print the slice split once, starting from the end, by numbers divisible by 3 (i.e. [50], [10, 40, 30, 20]):
fn main() { let v = [10, 40, 30, 20, 60, 50]; for group in v.rsplitn(1, |num| *num % 3 == 0) { println!("{:?}", group); } }

let v = [10, 40, 30, 20, 60, 50];
for group in v.rsplitn(1, |num| *num % 3 == 0) {
    println!("{:?}", group);
}

fn windows(&self, size: usize) -> Windows<Self::Item>

Returns an iterator over all contiguous windows of length size. The windows overlap. If the slice is shorter than size, the iterator returns no values.

Panics

Panics if size is 0.

Example

Print the adjacent pairs of a slice (i.e. [1,2], [2,3], [3,4]):
fn main() { let v = &[1, 2, 3, 4]; for win in v.windows(2) { println!("{:?}", win); } }

let v = &[1, 2, 3, 4];
for win in v.windows(2) {
    println!("{:?}", win);
}

fn chunks(&self, size: usize) -> Chunks<Self::Item>

Returns an iterator over size elements of the slice at a time. The chunks do not overlap. If size does not divide the length of the slice, then the last chunk will not have length size.

Panics

Panics if size is 0.

Example

Print the slice two elements at a time (i.e. [1,2], [3,4], [5]):
fn main() { let v = &[1, 2, 3, 4, 5]; for win in v.chunks(2) { println!("{:?}", win); } }

let v = &[1, 2, 3, 4, 5];
for win in v.chunks(2) {
    println!("{:?}", win);
}

fn get(&self, index: usize) -> Option<&Self::Item>

Returns the element of a slice at the given index, or None if the index is out of bounds.

Examples
fn main() { let v = [10, 40, 30]; assert_eq!(Some(&40), v.get(1)); assert_eq!(None, v.get(3)); }

let v = [10, 40, 30];
assert_eq!(Some(&40), v.get(1));
assert_eq!(None, v.get(3));

fn first(&self) -> Option<&Self::Item>

Returns the first element of a slice, or None if it is empty.

Examples
fn main() { let v = [10, 40, 30]; assert_eq!(Some(&10), v.first()); let w: &[i32] = &[]; assert_eq!(None, w.first()); }

let v = [10, 40, 30];
assert_eq!(Some(&10), v.first());

let w: &[i32] = &[];
assert_eq!(None, w.first());

fn tail(&self) -> &[Self::Item]

Returns all but the first element of a slice.

fn init(&self) -> &[Self::Item]

Returns all but the last element of a slice.

fn last(&self) -> Option<&Self::Item>

Returns the last element of a slice, or None if it is empty.

Examples
fn main() { let v = [10, 40, 30]; assert_eq!(Some(&30), v.last()); let w: &[i32] = &[]; assert_eq!(None, w.last()); }

let v = [10, 40, 30];
assert_eq!(Some(&30), v.last());

let w: &[i32] = &[];
assert_eq!(None, w.last());

unsafe fn get_unchecked(&self, index: usize) -> &Self::Item

Returns a pointer to the element at the given index, without doing bounds checking.

fn as_ptr(&self) -> *const Self::Item

Returns an unsafe pointer to the slice's buffer

The caller must ensure that the slice outlives the pointer this function returns, or else it will end up pointing to garbage.

Modifying the slice may cause its buffer to be reallocated, which would also make any pointers to it invalid.

fn binary_search_by<F>(&self, f: F) -> Result<usize, usize> where F: FnMut(&Self::Item) -> Ordering

Binary search a sorted slice with a comparator function.

The comparator function should implement an order consistent with the sort order of the underlying slice, returning an order code that indicates whether its argument is Less, Equal or Greater the desired target.

If a matching value is found then returns Ok, containing the index for the matched element; if no match is found then Err is returned, containing the index where a matching element could be inserted while maintaining sorted order.

Example

Looks up a series of four elements. The first is found, with a uniquely determined position; the second and third are not found; the fourth could match any position in [1,4].
fn main() { let s = [0, 1, 1, 1, 1, 2, 3, 5, 8, 13, 21, 34, 55]; let s = s.as_slice(); let seek = 13; assert_eq!(s.binary_search_by(|probe| probe.cmp(&seek)), Ok(9)); let seek = 4; assert_eq!(s.binary_search_by(|probe| probe.cmp(&seek)), Err(7)); let seek = 100; assert_eq!(s.binary_search_by(|probe| probe.cmp(&seek)), Err(13)); let seek = 1; let r = s.binary_search_by(|probe| probe.cmp(&seek)); assert!(match r { Ok(1...4) => true, _ => false, }); }

let s = [0, 1, 1, 1, 1, 2, 3, 5, 8, 13, 21, 34, 55];
let s = s.as_slice();

let seek = 13;
assert_eq!(s.binary_search_by(|probe| probe.cmp(&seek)), Ok(9));
let seek = 4;
assert_eq!(s.binary_search_by(|probe| probe.cmp(&seek)), Err(7));
let seek = 100;
assert_eq!(s.binary_search_by(|probe| probe.cmp(&seek)), Err(13));
let seek = 1;
let r = s.binary_search_by(|probe| probe.cmp(&seek));
assert!(match r { Ok(1...4) => true, _ => false, });

fn len(&self) -> usize

Return the number of elements in the slice

Example
fn main() { let a = [1, 2, 3]; assert_eq!(a.len(), 3); }

let a = [1, 2, 3];
assert_eq!(a.len(), 3);

fn get_mut(&mut self, index: usize) -> Option<&mut Self::Item>

Returns a mutable reference to the element at the given index, or None if the index is out of bounds

fn as_mut_slice(&mut self) -> &mut [Self::Item]

Work with self as a mut slice. Primarily intended for getting a &mut [T] from a [T; N].

fn slice_mut(&mut self, start: usize, end: usize) -> &mut [Self::Item]

Deprecated: use &mut s[start .. end] instead.

fn slice_from_mut(&mut self, start: usize) -> &mut [Self::Item]

Deprecated: use &mut s[start ..] instead.

fn slice_to_mut(&mut self, end: usize) -> &mut [Self::Item]

Deprecated: use &mut s[.. end] instead.

fn iter_mut(&mut self) -> IterMut<Self::Item>

Returns an iterator that allows modifying each value

fn first_mut(&mut self) -> Option<&mut Self::Item>

Returns a mutable pointer to the first element of a slice, or None if it is empty

fn tail_mut(&mut self) -> &mut [Self::Item]

Returns all but the first element of a mutable slice

fn init_mut(&mut self) -> &mut [Self::Item]

Returns all but the last element of a mutable slice

fn last_mut(&mut self) -> Option<&mut Self::Item>

Returns a mutable pointer to the last item in the slice.

fn split_mut<F>(&mut self, pred: F) -> SplitMut<Self::Item, F> where F: FnMut(&Self::Item) -> bool

Returns an iterator over mutable subslices separated by elements that match pred. The matched element is not contained in the subslices.

fn splitn_mut<F>(&mut self, n: usize, pred: F) -> SplitNMut<Self::Item, F> where F: FnMut(&Self::Item) -> bool

Returns an iterator over subslices separated by elements that match pred, limited to splitting at most n times. The matched element is not contained in the subslices.

fn rsplitn_mut<F>(&mut self, n: usize, pred: F) -> RSplitNMut<Self::Item, F> where F: FnMut(&Self::Item) -> bool

Returns an iterator over subslices separated by elements that match pred limited to splitting at most n times. This starts at the end of the slice and works backwards. The matched element is not contained in the subslices.

fn chunks_mut(&mut self, chunk_size: usize) -> ChunksMut<Self::Item>

Returns an iterator over chunk_size elements of the slice at a time. The chunks are mutable and do not overlap. If chunk_size does not divide the length of the slice, then the last chunk will not have length chunk_size.

Panics

Panics if chunk_size is 0.

fn swap(&mut self, a: usize, b: usize)

Swaps two elements in a slice.

Arguments

a - The index of the first element

b - The index of the second element

Panics

Panics if a or b are out of bounds.

Example
fn main() { let mut v = ["a", "b", "c", "d"]; v.swap(1, 3); assert!(v == ["a", "d", "c", "b"]); }

let mut v = ["a", "b", "c", "d"];
v.swap(1, 3);
assert!(v == ["a", "d", "c", "b"]);

fn split_at_mut(&mut self, mid: usize) -> (&mut [Self::Item], &mut [Self::Item])

Divides one &mut into two at an index.

The first will contain all indices from [0, mid) (excluding the index mid itself) and the second will contain all indices from [mid, len) (excluding the index len itself).

Panics

Panics if mid > len.

Example
fn main() { let mut v = [1, 2, 3, 4, 5, 6]; // scoped to restrict the lifetime of the borrows { let (left, right) = v.split_at_mut(0); assert!(left == []); assert!(right == [1, 2, 3, 4, 5, 6]); } { let (left, right) = v.split_at_mut(2); assert!(left == [1, 2]); assert!(right == [3, 4, 5, 6]); } { let (left, right) = v.split_at_mut(6); assert!(left == [1, 2, 3, 4, 5, 6]); assert!(right == []); } }

let mut v = [1, 2, 3, 4, 5, 6];

// scoped to restrict the lifetime of the borrows
{
   let (left, right) = v.split_at_mut(0);
   assert!(left == []);
   assert!(right == [1, 2, 3, 4, 5, 6]);
}

{
    let (left, right) = v.split_at_mut(2);
    assert!(left == [1, 2]);
    assert!(right == [3, 4, 5, 6]);
}

{
    let (left, right) = v.split_at_mut(6);
    assert!(left == [1, 2, 3, 4, 5, 6]);
    assert!(right == []);
}

fn reverse(&mut self)

Reverse the order of elements in a slice, in place.

Example
fn main() { let mut v = [1, 2, 3]; v.reverse(); assert!(v == [3, 2, 1]); }

let mut v = [1, 2, 3];
v.reverse();
assert!(v == [3, 2, 1]);

unsafe fn get_unchecked_mut(&mut self, index: usize) -> &mut Self::Item

Returns an unsafe mutable pointer to the element in index

fn as_mut_ptr(&mut self) -> *mut Self::Item

Return an unsafe mutable pointer to the slice's buffer.

The caller must ensure that the slice outlives the pointer this function returns, or else it will end up pointing to garbage.

Modifying the slice may cause its buffer to be reallocated, which would also make any pointers to it invalid.

fn to_vec(&self) -> Vec<Self::Item> where Self::Item: Clone

Copies self into a new Vec.

fn permutations(&self) -> Permutations<Self::Item> where Self::Item: Clone

Creates an iterator that yields every possible permutation of the vector in succession.

Examples
fn main() { let v = [1, 2, 3]; let mut perms = v.permutations(); for p in perms { println!("{:?}", p); } }

let v = [1, 2, 3];
let mut perms = v.permutations();

for p in perms {
  println!("{:?}", p);
}

Iterating through permutations one by one.
fn main() { let v = [1, 2, 3]; let mut perms = v.permutations(); assert_eq!(Some(vec![1, 2, 3]), perms.next()); assert_eq!(Some(vec![1, 3, 2]), perms.next()); assert_eq!(Some(vec![3, 1, 2]), perms.next()); }

let v = [1, 2, 3];
let mut perms = v.permutations();

assert_eq!(Some(vec![1, 2, 3]), perms.next());
assert_eq!(Some(vec![1, 3, 2]), perms.next());
assert_eq!(Some(vec![3, 1, 2]), perms.next());

fn clone_from_slice(&mut self, &[Self::Item]) -> usize where Self::Item: Clone

Copies as many elements from src as it can into self (the shorter of self.len() and src.len()). Returns the number of elements copied.

Example
fn main() { let mut dst = [0, 0, 0]; let src = [1, 2]; assert!(dst.clone_from_slice(&src) == 2); assert!(dst == [1, 2, 0]); let src2 = [3, 4, 5, 6]; assert!(dst.clone_from_slice(&src2) == 3); assert!(dst == [3, 4, 5]); }

let mut dst = [0, 0, 0];
let src = [1, 2];

assert!(dst.clone_from_slice(&src) == 2);
assert!(dst == [1, 2, 0]);

let src2 = [3, 4, 5, 6];
assert!(dst.clone_from_slice(&src2) == 3);
assert!(dst == [3, 4, 5]);

fn sort(&mut self) where Self::Item: Ord

Sorts the slice, in place.

This is equivalent to self.sort_by(|a, b| a.cmp(b)).

Examples
fn main() { let mut v = [-5, 4, 1, -3, 2]; v.sort(); assert!(v == [-5, -3, 1, 2, 4]); }

let mut v = [-5, 4, 1, -3, 2];

v.sort();
assert!(v == [-5, -3, 1, 2, 4]);

fn binary_search(&self, x: &Self::Item) -> Result<usize, usize> where Self::Item: Ord

Binary search a sorted slice for a given element.

If the value is found then Ok is returned, containing the index of the matching element; if the value is not found then Err is returned, containing the index where a matching element could be inserted while maintaining sorted order.

Example

Looks up a series of four elements. The first is found, with a uniquely determined position; the second and third are not found; the fourth could match any position in [1,4].
fn main() { let s = [0, 1, 1, 1, 1, 2, 3, 5, 8, 13, 21, 34, 55]; let s = s.as_slice(); assert_eq!(s.binary_search(&13), Ok(9)); assert_eq!(s.binary_search(&4), Err(7)); assert_eq!(s.binary_search(&100), Err(13)); let r = s.binary_search(&1); assert!(match r { Ok(1...4) => true, _ => false, }); }

let s = [0, 1, 1, 1, 1, 2, 3, 5, 8, 13, 21, 34, 55];
let s = s.as_slice();

assert_eq!(s.binary_search(&13),  Ok(9));
assert_eq!(s.binary_search(&4),   Err(7));
assert_eq!(s.binary_search(&100), Err(13));
let r = s.binary_search(&1);
assert!(match r { Ok(1...4) => true, _ => false, });

fn next_permutation(&mut self) -> bool where Self::Item: Ord

Mutates the slice to the next lexicographic permutation.

Returns true if successful and false if the slice is at the last-ordered permutation.

Example
fn main() { let v: &mut [_] = &mut [0, 1, 2]; v.next_permutation(); let b: &mut [_] = &mut [0, 2, 1]; assert!(v == b); v.next_permutation(); let b: &mut [_] = &mut [1, 0, 2]; assert!(v == b); }

let v: &mut [_] = &mut [0, 1, 2];
v.next_permutation();
let b: &mut [_] = &mut [0, 2, 1];
assert!(v == b);
v.next_permutation();
let b: &mut [_] = &mut [1, 0, 2];
assert!(v == b);

fn prev_permutation(&mut self) -> bool where Self::Item: Ord

Mutates the slice to the previous lexicographic permutation.

Returns true if successful and false if the slice is at the first-ordered permutation.

Example
fn main() { let v: &mut [_] = &mut [1, 0, 2]; v.prev_permutation(); let b: &mut [_] = &mut [0, 2, 1]; assert!(v == b); v.prev_permutation(); let b: &mut [_] = &mut [0, 1, 2]; assert!(v == b); }

let v: &mut [_] = &mut [1, 0, 2];
v.prev_permutation();
let b: &mut [_] = &mut [0, 2, 1];
assert!(v == b);
v.prev_permutation();
let b: &mut [_] = &mut [0, 1, 2];
assert!(v == b);

fn position_elem(&self, t: &Self::Item) -> Option<usize> where Self::Item: PartialEq

Find the first index containing a matching value.

fn rposition_elem(&self, t: &Self::Item) -> Option<usize> where Self::Item: PartialEq

Find the last index containing a matching value.

fn contains(&self, x: &Self::Item) -> bool where Self::Item: PartialEq

Returns true if the slice contains an element with the given value.

Examples
fn main() { let v = [10, 40, 30]; assert!(v.contains(&30)); assert!(!v.contains(&50)); }

let v = [10, 40, 30];
assert!(v.contains(&30));
assert!(!v.contains(&50));

fn starts_with(&self, needle: &[Self::Item]) -> bool where Self::Item: PartialEq

Returns true if needle is a prefix of the slice.

Examples
fn main() { let v = [10, 40, 30]; assert!(v.starts_with(&[10])); assert!(v.starts_with(&[10, 40])); assert!(!v.starts_with(&[50])); assert!(!v.starts_with(&[10, 50])); }

let v = [10, 40, 30];
assert!(v.starts_with(&[10]));
assert!(v.starts_with(&[10, 40]));
assert!(!v.starts_with(&[50]));
assert!(!v.starts_with(&[10, 50]));

fn ends_with(&self, needle: &[Self::Item]) -> bool where Self::Item: PartialEq

Returns true if needle is a suffix of the slice.

Examples
fn main() { let v = [10, 40, 30]; assert!(v.ends_with(&[30])); assert!(v.ends_with(&[40, 30])); assert!(!v.ends_with(&[50])); assert!(!v.ends_with(&[50, 30])); }

let v = [10, 40, 30];
assert!(v.ends_with(&[30]));
assert!(v.ends_with(&[40, 30]));
assert!(!v.ends_with(&[50]));
assert!(!v.ends_with(&[50, 30]));

fn into_vec(self: Box<Self>) -> Vec<Self::Item>

Convert self into a vector without clones or allocation.

Provided Methods

fn is_empty(&self) -> bool

Returns true if the slice has a length of 0

Example
fn main() { let a = [1, 2, 3]; assert!(!a.is_empty()); }

let a = [1, 2, 3];
assert!(!a.is_empty());

fn binary_search_elem(&self, x: &Self::Item) -> Result<usize, usize> where Self::Item: Ord

Deprecated: use binary_search instead.