Struct collections::str::CharIndices

pub struct CharIndices<'a> { /* fields omitted */ }

An iterator over the chars of a string slice, and their positions.

This struct is created by the char_indices method on str. See its documentation for more.

Methods

impl<'a> CharIndices<'a>

fn as_str(&self) -> &'a str

View the underlying data as a subslice of the original data.

This has the same lifetime as the original slice, and so the iterator can continue to be used while this exists.