Struct collections::str::Chars

pub struct Chars<'a> { /* fields omitted */ }

An iterator over the chars of a string slice.

This struct is created by the chars method on str. See its documentation for more.

Methods

impl<'a> Chars<'a>

fn as_str(&self) -> &'a str

View the underlying data as a subslice of the original data.

This has the same lifetime as the original slice, and so the iterator can continue to be used while this exists.

Examples

let mut chars = "abc".chars();

assert_eq!(chars.as_str(), "abc");
chars.next();
assert_eq!(chars.as_str(), "bc");
chars.next();
chars.next();
assert_eq!(chars.as_str(), "");