Struct collections::str::Utf8Error

pub struct Utf8Error {
    // some fields omitted
}

Errors which can occur when attempting to interpret a byte slice as a str.

Methods

impl Utf8Error

fn valid_up_to(&self) -> usize

Unstable
: method just added

Returns the index in the given string up to which valid UTF-8 was verified.

Starting at the index provided, but not necessarily at it precisely, an invalid UTF-8 encoding sequence was found.