Trait collections::str::CharEqDeprecated

pub trait CharEq {
    fn matches(&mut self, char) -> bool;
    fn only_ascii(&self) -> bool;
}

Something that can be used to compare against a character

Required Methods

fn matches(&mut self, char) -> bool

Determine if the splitter should split at the given character

fn only_ascii(&self) -> bool

Indicate if this is only concerned about ASCII characters, which can allow for a faster implementation.