Trait collections::string::ToString

pub trait ToString {
    fn to_string(&self) -> String;
}

A trait for converting a value to a String.

This trait is automatically implemented for any type which implements the Display trait. As such, ToString shouldn't be implemented directly: Display should be implemented instead, and you get the ToString implementation for free.

Required Methods

fn to_string(&self) -> String

Converts the given value to a String.

Examples

Basic usage:

let i = 5;
let five = String::from("5");

assert_eq!(five, i.to_string());Run