Struct collections::BTreeSetStable

pub struct BTreeSet<T> {
    // some fields omitted
}

A set based on a B-Tree.

See BTreeMap's documentation for a detailed discussion of this collection's performance benefits and drawbacks.

Methods

impl<T: Ord> BTreeSet<T>

fn new() -> BTreeSet<T>

Makes a new BTreeSet with a reasonable choice of B.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let mut set: BTreeSet<i32> = BTreeSet::new(); }

use std::collections::BTreeSet;

let mut set: BTreeSet<i32> = BTreeSet::new();

fn with_b(b: usize) -> BTreeSet<T>

Makes a new BTreeSet with the given B.

B cannot be less than 2.

impl<T> BTreeSet<T>

fn iter(&self) -> Iter<T>

Gets an iterator over the BTreeSet's contents.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let set: BTreeSet<usize> = [1, 2, 3, 4].iter().cloned().collect(); for x in set.iter() { println!("{}", x); } let v: Vec<usize> = set.iter().cloned().collect(); assert_eq!(v, [1, 2, 3, 4]); }

use std::collections::BTreeSet;

let set: BTreeSet<usize> = [1, 2, 3, 4].iter().cloned().collect();

for x in set.iter() {
    println!("{}", x);
}

let v: Vec<usize> = set.iter().cloned().collect();
assert_eq!(v, [1, 2, 3, 4]);

fn into_iter(self) -> IntoIter<T>

Gets an iterator for moving out the BtreeSet's contents.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let set: BTreeSet<usize> = [1, 2, 3, 4].iter().cloned().collect(); let v: Vec<usize> = set.into_iter().collect(); assert_eq!(v, [1, 2, 3, 4]); }

use std::collections::BTreeSet;

let set: BTreeSet<usize> = [1, 2, 3, 4].iter().cloned().collect();

let v: Vec<usize> = set.into_iter().collect();
assert_eq!(v, [1, 2, 3, 4]);

impl<T: Ord> BTreeSet<T>

fn range<'a>(&'a self, min: Bound<&T>, max: Bound<&T>) -> Range<'a, T>

Constructs a double-ended iterator over a sub-range of elements in the set, starting at min, and ending at max. If min is Unbounded, then it will be treated as "negative infinity", and if max is Unbounded, then it will be treated as "positive infinity". Thus range(Unbounded, Unbounded) will yield the whole collection.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; use std::collections::Bound::{Included, Unbounded}; let mut set = BTreeSet::new(); set.insert(3); set.insert(5); set.insert(8); for &elem in set.range(Included(&4), Included(&8)) { println!("{}", elem); } assert_eq!(Some(&5), set.range(Included(&4), Unbounded).next()); }

use std::collections::BTreeSet;
use std::collections::Bound::{Included, Unbounded};

let mut set = BTreeSet::new();
set.insert(3);
set.insert(5);
set.insert(8);
for &elem in set.range(Included(&4), Included(&8)) {
    println!("{}", elem);
}
assert_eq!(Some(&5), set.range(Included(&4), Unbounded).next());

impl<T: Ord> BTreeSet<T>

fn difference<'a>(&'a self, other: &'a BTreeSet<T>) -> Difference<'a, T>

Visits the values representing the difference, in ascending order.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let mut a = BTreeSet::new(); a.insert(1); a.insert(2); let mut b = BTreeSet::new(); b.insert(2); b.insert(3); let diff: Vec<usize> = a.difference(&b).cloned().collect(); assert_eq!(diff, [1]); }

use std::collections::BTreeSet;

let mut a = BTreeSet::new();
a.insert(1);
a.insert(2);

let mut b = BTreeSet::new();
b.insert(2);
b.insert(3);

let diff: Vec<usize> = a.difference(&b).cloned().collect();
assert_eq!(diff, [1]);

fn symmetric_difference<'a>(&'a self, other: &'a BTreeSet<T>) -> SymmetricDifference<'a, T>

Visits the values representing the symmetric difference, in ascending order.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let mut a = BTreeSet::new(); a.insert(1); a.insert(2); let mut b = BTreeSet::new(); b.insert(2); b.insert(3); let sym_diff: Vec<usize> = a.symmetric_difference(&b).cloned().collect(); assert_eq!(sym_diff, [1, 3]); }

use std::collections::BTreeSet;

let mut a = BTreeSet::new();
a.insert(1);
a.insert(2);

let mut b = BTreeSet::new();
b.insert(2);
b.insert(3);

let sym_diff: Vec<usize> = a.symmetric_difference(&b).cloned().collect();
assert_eq!(sym_diff, [1, 3]);

fn intersection<'a>(&'a self, other: &'a BTreeSet<T>) -> Intersection<'a, T>

Visits the values representing the intersection, in ascending order.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let mut a = BTreeSet::new(); a.insert(1); a.insert(2); let mut b = BTreeSet::new(); b.insert(2); b.insert(3); let intersection: Vec<usize> = a.intersection(&b).cloned().collect(); assert_eq!(intersection, [2]); }

use std::collections::BTreeSet;

let mut a = BTreeSet::new();
a.insert(1);
a.insert(2);

let mut b = BTreeSet::new();
b.insert(2);
b.insert(3);

let intersection: Vec<usize> = a.intersection(&b).cloned().collect();
assert_eq!(intersection, [2]);

fn union<'a>(&'a self, other: &'a BTreeSet<T>) -> Union<'a, T>

Visits the values representing the union, in ascending order.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let mut a = BTreeSet::new(); a.insert(1); let mut b = BTreeSet::new(); b.insert(2); let union: Vec<usize> = a.union(&b).cloned().collect(); assert_eq!(union, [1, 2]); }

use std::collections::BTreeSet;

let mut a = BTreeSet::new();
a.insert(1);

let mut b = BTreeSet::new();
b.insert(2);

let union: Vec<usize> = a.union(&b).cloned().collect();
assert_eq!(union, [1, 2]);

fn len(&self) -> usize

Return the number of elements in the set

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let mut v = BTreeSet::new(); assert_eq!(v.len(), 0); v.insert(1); assert_eq!(v.len(), 1); }

use std::collections::BTreeSet;

let mut v = BTreeSet::new();
assert_eq!(v.len(), 0);
v.insert(1);
assert_eq!(v.len(), 1);

fn is_empty(&self) -> bool

Returns true if the set contains no elements

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let mut v = BTreeSet::new(); assert!(v.is_empty()); v.insert(1); assert!(!v.is_empty()); }

use std::collections::BTreeSet;

let mut v = BTreeSet::new();
assert!(v.is_empty());
v.insert(1);
assert!(!v.is_empty());

fn clear(&mut self)

Clears the set, removing all values.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let mut v = BTreeSet::new(); v.insert(1); v.clear(); assert!(v.is_empty()); }

use std::collections::BTreeSet;

let mut v = BTreeSet::new();
v.insert(1);
v.clear();
assert!(v.is_empty());

fn contains<Q: ?Sized>(&self, value: &Q) -> bool where T: Borrow<Q>, Q: Ord

Returns true if the set contains a value.

The value may be any borrowed form of the set's value type, but the ordering on the borrowed form must match the ordering on the value type.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let set: BTreeSet<_> = [1, 2, 3].iter().cloned().collect(); assert_eq!(set.contains(&1), true); assert_eq!(set.contains(&4), false); }

use std::collections::BTreeSet;

let set: BTreeSet<_> = [1, 2, 3].iter().cloned().collect();
assert_eq!(set.contains(&1), true);
assert_eq!(set.contains(&4), false);

fn is_disjoint(&self, other: &BTreeSet<T>) -> bool

Returns true if the set has no elements in common with other. This is equivalent to checking for an empty intersection.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let a: BTreeSet<_> = [1, 2, 3].iter().cloned().collect(); let mut b = BTreeSet::new(); assert_eq!(a.is_disjoint(&b), true); b.insert(4); assert_eq!(a.is_disjoint(&b), true); b.insert(1); assert_eq!(a.is_disjoint(&b), false); }

use std::collections::BTreeSet;

let a: BTreeSet<_> = [1, 2, 3].iter().cloned().collect();
let mut b = BTreeSet::new();

assert_eq!(a.is_disjoint(&b), true);
b.insert(4);
assert_eq!(a.is_disjoint(&b), true);
b.insert(1);
assert_eq!(a.is_disjoint(&b), false);

fn is_subset(&self, other: &BTreeSet<T>) -> bool

Returns true if the set is a subset of another.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let sup: BTreeSet<_> = [1, 2, 3].iter().cloned().collect(); let mut set = BTreeSet::new(); assert_eq!(set.is_subset(&sup), true); set.insert(2); assert_eq!(set.is_subset(&sup), true); set.insert(4); assert_eq!(set.is_subset(&sup), false); }

use std::collections::BTreeSet;

let sup: BTreeSet<_> = [1, 2, 3].iter().cloned().collect();
let mut set = BTreeSet::new();

assert_eq!(set.is_subset(&sup), true);
set.insert(2);
assert_eq!(set.is_subset(&sup), true);
set.insert(4);
assert_eq!(set.is_subset(&sup), false);

fn is_superset(&self, other: &BTreeSet<T>) -> bool

Returns true if the set is a superset of another.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let sub: BTreeSet<_> = [1, 2].iter().cloned().collect(); let mut set = BTreeSet::new(); assert_eq!(set.is_superset(&sub), false); set.insert(0); set.insert(1); assert_eq!(set.is_superset(&sub), false); set.insert(2); assert_eq!(set.is_superset(&sub), true); }

use std::collections::BTreeSet;

let sub: BTreeSet<_> = [1, 2].iter().cloned().collect();
let mut set = BTreeSet::new();

assert_eq!(set.is_superset(&sub), false);

set.insert(0);
set.insert(1);
assert_eq!(set.is_superset(&sub), false);

set.insert(2);
assert_eq!(set.is_superset(&sub), true);

fn insert(&mut self, value: T) -> bool

Adds a value to the set. Returns true if the value was not already present in the set.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let mut set = BTreeSet::new(); assert_eq!(set.insert(2), true); assert_eq!(set.insert(2), false); assert_eq!(set.len(), 1); }

use std::collections::BTreeSet;

let mut set = BTreeSet::new();

assert_eq!(set.insert(2), true);
assert_eq!(set.insert(2), false);
assert_eq!(set.len(), 1);

fn remove<Q: ?Sized>(&mut self, value: &Q) -> bool where T: Borrow<Q>, Q: Ord

Removes a value from the set. Returns true if the value was present in the set.

The value may be any borrowed form of the set's value type, but the ordering on the borrowed form must match the ordering on the value type.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let mut set = BTreeSet::new(); set.insert(2); assert_eq!(set.remove(&2), true); assert_eq!(set.remove(&2), false); }

use std::collections::BTreeSet;

let mut set = BTreeSet::new();

set.insert(2);
assert_eq!(set.remove(&2), true);
assert_eq!(set.remove(&2), false);

impl<T: Ord> BTreeSet<T>

fn new() -> BTreeSet<T>

Makes a new BTreeSet with a reasonable choice of B.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let mut set: BTreeSet<i32> = BTreeSet::new(); }

use std::collections::BTreeSet;

let mut set: BTreeSet<i32> = BTreeSet::new();

fn with_b(b: usize) -> BTreeSet<T>

Makes a new BTreeSet with the given B.

B cannot be less than 2.

impl<T> BTreeSet<T>

fn iter(&self) -> Iter<T>

Gets an iterator over the BTreeSet's contents.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let set: BTreeSet<usize> = [1, 2, 3, 4].iter().cloned().collect(); for x in set.iter() { println!("{}", x); } let v: Vec<usize> = set.iter().cloned().collect(); assert_eq!(v, [1, 2, 3, 4]); }

use std::collections::BTreeSet;

let set: BTreeSet<usize> = [1, 2, 3, 4].iter().cloned().collect();

for x in set.iter() {
    println!("{}", x);
}

let v: Vec<usize> = set.iter().cloned().collect();
assert_eq!(v, [1, 2, 3, 4]);

fn into_iter(self) -> IntoIter<T>

Gets an iterator for moving out the BtreeSet's contents.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let set: BTreeSet<usize> = [1, 2, 3, 4].iter().cloned().collect(); let v: Vec<usize> = set.into_iter().collect(); assert_eq!(v, [1, 2, 3, 4]); }

use std::collections::BTreeSet;

let set: BTreeSet<usize> = [1, 2, 3, 4].iter().cloned().collect();

let v: Vec<usize> = set.into_iter().collect();
assert_eq!(v, [1, 2, 3, 4]);

impl<T: Ord> BTreeSet<T>

fn range<'a>(&'a self, min: Bound<&T>, max: Bound<&T>) -> Range<'a, T>

Constructs a double-ended iterator over a sub-range of elements in the set, starting at min, and ending at max. If min is Unbounded, then it will be treated as "negative infinity", and if max is Unbounded, then it will be treated as "positive infinity". Thus range(Unbounded, Unbounded) will yield the whole collection.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; use std::collections::Bound::{Included, Unbounded}; let mut set = BTreeSet::new(); set.insert(3); set.insert(5); set.insert(8); for &elem in set.range(Included(&4), Included(&8)) { println!("{}", elem); } assert_eq!(Some(&5), set.range(Included(&4), Unbounded).next()); }

use std::collections::BTreeSet;
use std::collections::Bound::{Included, Unbounded};

let mut set = BTreeSet::new();
set.insert(3);
set.insert(5);
set.insert(8);
for &elem in set.range(Included(&4), Included(&8)) {
    println!("{}", elem);
}
assert_eq!(Some(&5), set.range(Included(&4), Unbounded).next());

impl<T: Ord> BTreeSet<T>

fn difference<'a>(&'a self, other: &'a BTreeSet<T>) -> Difference<'a, T>

Visits the values representing the difference, in ascending order.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let mut a = BTreeSet::new(); a.insert(1); a.insert(2); let mut b = BTreeSet::new(); b.insert(2); b.insert(3); let diff: Vec<usize> = a.difference(&b).cloned().collect(); assert_eq!(diff, [1]); }

use std::collections::BTreeSet;

let mut a = BTreeSet::new();
a.insert(1);
a.insert(2);

let mut b = BTreeSet::new();
b.insert(2);
b.insert(3);

let diff: Vec<usize> = a.difference(&b).cloned().collect();
assert_eq!(diff, [1]);

fn symmetric_difference<'a>(&'a self, other: &'a BTreeSet<T>) -> SymmetricDifference<'a, T>

Visits the values representing the symmetric difference, in ascending order.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let mut a = BTreeSet::new(); a.insert(1); a.insert(2); let mut b = BTreeSet::new(); b.insert(2); b.insert(3); let sym_diff: Vec<usize> = a.symmetric_difference(&b).cloned().collect(); assert_eq!(sym_diff, [1, 3]); }

use std::collections::BTreeSet;

let mut a = BTreeSet::new();
a.insert(1);
a.insert(2);

let mut b = BTreeSet::new();
b.insert(2);
b.insert(3);

let sym_diff: Vec<usize> = a.symmetric_difference(&b).cloned().collect();
assert_eq!(sym_diff, [1, 3]);

fn intersection<'a>(&'a self, other: &'a BTreeSet<T>) -> Intersection<'a, T>

Visits the values representing the intersection, in ascending order.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let mut a = BTreeSet::new(); a.insert(1); a.insert(2); let mut b = BTreeSet::new(); b.insert(2); b.insert(3); let intersection: Vec<usize> = a.intersection(&b).cloned().collect(); assert_eq!(intersection, [2]); }

use std::collections::BTreeSet;

let mut a = BTreeSet::new();
a.insert(1);
a.insert(2);

let mut b = BTreeSet::new();
b.insert(2);
b.insert(3);

let intersection: Vec<usize> = a.intersection(&b).cloned().collect();
assert_eq!(intersection, [2]);

fn union<'a>(&'a self, other: &'a BTreeSet<T>) -> Union<'a, T>

Visits the values representing the union, in ascending order.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let mut a = BTreeSet::new(); a.insert(1); let mut b = BTreeSet::new(); b.insert(2); let union: Vec<usize> = a.union(&b).cloned().collect(); assert_eq!(union, [1, 2]); }

use std::collections::BTreeSet;

let mut a = BTreeSet::new();
a.insert(1);

let mut b = BTreeSet::new();
b.insert(2);

let union: Vec<usize> = a.union(&b).cloned().collect();
assert_eq!(union, [1, 2]);

fn len(&self) -> usize

Return the number of elements in the set

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let mut v = BTreeSet::new(); assert_eq!(v.len(), 0); v.insert(1); assert_eq!(v.len(), 1); }

use std::collections::BTreeSet;

let mut v = BTreeSet::new();
assert_eq!(v.len(), 0);
v.insert(1);
assert_eq!(v.len(), 1);

fn is_empty(&self) -> bool

Returns true if the set contains no elements

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let mut v = BTreeSet::new(); assert!(v.is_empty()); v.insert(1); assert!(!v.is_empty()); }

use std::collections::BTreeSet;

let mut v = BTreeSet::new();
assert!(v.is_empty());
v.insert(1);
assert!(!v.is_empty());

fn clear(&mut self)

Clears the set, removing all values.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let mut v = BTreeSet::new(); v.insert(1); v.clear(); assert!(v.is_empty()); }

use std::collections::BTreeSet;

let mut v = BTreeSet::new();
v.insert(1);
v.clear();
assert!(v.is_empty());

fn contains<Q: ?Sized>(&self, value: &Q) -> bool where T: Borrow<Q>, Q: Ord

Returns true if the set contains a value.

The value may be any borrowed form of the set's value type, but the ordering on the borrowed form must match the ordering on the value type.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let set: BTreeSet<_> = [1, 2, 3].iter().cloned().collect(); assert_eq!(set.contains(&1), true); assert_eq!(set.contains(&4), false); }

use std::collections::BTreeSet;

let set: BTreeSet<_> = [1, 2, 3].iter().cloned().collect();
assert_eq!(set.contains(&1), true);
assert_eq!(set.contains(&4), false);

fn is_disjoint(&self, other: &BTreeSet<T>) -> bool

Returns true if the set has no elements in common with other. This is equivalent to checking for an empty intersection.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let a: BTreeSet<_> = [1, 2, 3].iter().cloned().collect(); let mut b = BTreeSet::new(); assert_eq!(a.is_disjoint(&b), true); b.insert(4); assert_eq!(a.is_disjoint(&b), true); b.insert(1); assert_eq!(a.is_disjoint(&b), false); }

use std::collections::BTreeSet;

let a: BTreeSet<_> = [1, 2, 3].iter().cloned().collect();
let mut b = BTreeSet::new();

assert_eq!(a.is_disjoint(&b), true);
b.insert(4);
assert_eq!(a.is_disjoint(&b), true);
b.insert(1);
assert_eq!(a.is_disjoint(&b), false);

fn is_subset(&self, other: &BTreeSet<T>) -> bool

Returns true if the set is a subset of another.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let sup: BTreeSet<_> = [1, 2, 3].iter().cloned().collect(); let mut set = BTreeSet::new(); assert_eq!(set.is_subset(&sup), true); set.insert(2); assert_eq!(set.is_subset(&sup), true); set.insert(4); assert_eq!(set.is_subset(&sup), false); }

use std::collections::BTreeSet;

let sup: BTreeSet<_> = [1, 2, 3].iter().cloned().collect();
let mut set = BTreeSet::new();

assert_eq!(set.is_subset(&sup), true);
set.insert(2);
assert_eq!(set.is_subset(&sup), true);
set.insert(4);
assert_eq!(set.is_subset(&sup), false);

fn is_superset(&self, other: &BTreeSet<T>) -> bool

Returns true if the set is a superset of another.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let sub: BTreeSet<_> = [1, 2].iter().cloned().collect(); let mut set = BTreeSet::new(); assert_eq!(set.is_superset(&sub), false); set.insert(0); set.insert(1); assert_eq!(set.is_superset(&sub), false); set.insert(2); assert_eq!(set.is_superset(&sub), true); }

use std::collections::BTreeSet;

let sub: BTreeSet<_> = [1, 2].iter().cloned().collect();
let mut set = BTreeSet::new();

assert_eq!(set.is_superset(&sub), false);

set.insert(0);
set.insert(1);
assert_eq!(set.is_superset(&sub), false);

set.insert(2);
assert_eq!(set.is_superset(&sub), true);

fn insert(&mut self, value: T) -> bool

Adds a value to the set. Returns true if the value was not already present in the set.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let mut set = BTreeSet::new(); assert_eq!(set.insert(2), true); assert_eq!(set.insert(2), false); assert_eq!(set.len(), 1); }

use std::collections::BTreeSet;

let mut set = BTreeSet::new();

assert_eq!(set.insert(2), true);
assert_eq!(set.insert(2), false);
assert_eq!(set.len(), 1);

fn remove<Q: ?Sized>(&mut self, value: &Q) -> bool where T: Borrow<Q>, Q: Ord

Removes a value from the set. Returns true if the value was present in the set.

The value may be any borrowed form of the set's value type, but the ordering on the borrowed form must match the ordering on the value type.

Examples
extern crate collections; fn main() { use std::collections::BTreeSet; let mut set = BTreeSet::new(); set.insert(2); assert_eq!(set.remove(&2), true); assert_eq!(set.remove(&2), false); }

use std::collections::BTreeSet;

let mut set = BTreeSet::new();

set.insert(2);
assert_eq!(set.remove(&2), true);
assert_eq!(set.remove(&2), false);