Module collections::vec_mapUnstable

A simple map based on a vector for small integer keys. Space requirements are O(highest integer key).

Structs

Drain

IntoIter
A consuming iterator over the key-value pairs of a map.

Iter
An iterator over the key-value pairs of a map.

IterMut
An iterator over the key-value pairs of a map, with the values being mutable.

Keys
An iterator over the keys of a map.

Values
An iterator over the values of a map.

VecMap
A map optimized for small integer keys.