Enum collections::vec_map::EntryUnstable

pub enum Entry<'a, V: 'a> {
    Vacant(VacantEntry<'a, V>),
    Occupied(OccupiedEntry<'a, V>),
}

A view into a single entry in a map, which may either be vacant or occupied.

Variants

Vacant

A vacant Entry

Occupied

An occupied Entry

Methods

impl<'a, V> Entry<'a, V>

fn get(self) -> Result<&'a mut V, VacantEntry<'a, V>>

Returns a mutable reference to the entry if occupied, or the VacantEntry if vacant