Crate collections

Unstable (collections #27783)
: library is unlikely to be stabilized with the current layout and name, use std::collections instead

Collection types.

See std::collections for a detailed discussion of collections in Rust.

Reexports

pub use binary_heap::BinaryHeap;

pub use linked_list::LinkedList;

pub use enum_set::EnumSet;

pub use vec_deque::VecDeque;

pub use string::String;

pub use vec::Vec;

Modules

binary_heap
A priority queue implemented with a binary heap.

borrow
A module for working with borrowed data.

boxed
A pointer type for heap allocation.

btree_map

btree_set

fmt
Utilities for formatting and printing strings

linked_list
A doubly-linked list with owned nodes.

slice
A dynamically-sized view into a contiguous sequence, [T].

str
Unicode string slices

string
A UTF-8 encoded, growable string.

vec
A growable list type with heap-allocated contents, written Vec<T> but pronounced 'vector.'

vec_deque
VecDeque is a double-ended queue, which is implemented with the help of a growing ring buffer.

enum_set 	 [Unstable]
A structure for holding a set of enum variants.

range 	 [Unstable]
Range syntax.

Macros

format!
Use the syntax described in std::fmt to create a value of type String. See std::fmt for more information.

vec!
Creates a Vec containing the arguments.

Structs

BTreeMap
A map based on a B-Tree.

BTreeSet
A set based on a B-Tree.

Enums

Bound 	 [Unstable]
An endpoint of a range of keys.