Module collections::btree_set

Reexports

use core::prelude::v1::*;

use core::prelude::v1::*;

use core::cmp::Ordering::{self, Less, Greater, Equal};

use core::cmp::{min, max};

use core::fmt::Debug;

use core::fmt;

use core::iter::{Peekable, FromIterator};

use core::ops::{BitOr, BitAnd, BitXor, Sub};

use borrow::Borrow;

use btree_map::{BTreeMap, Keys};

use super::Recover;

use Bound;

Structs

BTreeSet
A set based on a B-Tree.

Difference
A lazy iterator producing elements in the set difference (in-order).

Intersection
A lazy iterator producing elements in the set intersection (in-order).

IntoIter
An owning iterator over a BTreeSet's items.

Iter
An iterator over a BTreeSet's items.

SymmetricDifference
A lazy iterator producing elements in the set symmetric difference (in-order).

Union
A lazy iterator producing elements in the set union (in-order).

Range 	 [Unstable]
An iterator over a sub-range of BTreeSet's items.

Functions

cmp_opt 	 [Unstable]
Compare x and y, but return short if x is None and long if y is None