Module collections::stringStable

An owned, growable string that enforces that its contents are valid UTF-8.

Structs

DerefString
Wrapper type providing a &String reference via Deref.

FromUtf16Error
A possible error value from the String::from_utf16 function.

FromUtf8Error
A possible error value from the String::from_utf8 function.

String
A growable string stored as a UTF-8 encoded buffer.

Traits

ToString
A generic trait for converting a value to a string

Functions

as_string
Convert a string slice to a wrapper type providing a &String reference.

Type Definitions

CowString
A clone-on-write string